+++ front/src/app/modules/sessions-form/interfaces/admin-session.ts
@@ -7,6 +7,7 @@ export interface IAdminSession extends IAdminUpdateSession {
 
 export interface IAdminUpdateSession {
   id: number;
+  title: string;
   startDateTime: number;
   endDateTime: number;
   blockContainerCreationBeforeStartTime: boolean;